add support for multiple vstubs
Multiple vstubs can be serviced via a single vstub-server executable.
The concept of vstub module is introduced and framework routines call
per-module handler routines.
Currently 7 vstub modules are supported. Each module is assigned with
code as follows:
- hid-mouse: Mouse
- hid-keyboard: Keyboard
- cdc: CDC ACM
- bth: ASUS blutooth 2.0
- cp210x: CP210x USB to UART Bridge VCP Drivers - Silicon Labs
- arduino: Arduino Uno
- avrmkii: AVRISP MK II
If vstub-server start with no argument, it starts with all the vstubs
attached. The server can start with a specific vstub by providing
a vstub code as an argument.